feat: add pagination and sorting to organizations admin page
- Add pagination support to adminGetAll tRPC query with ZAdminGetAllInput schema
- Replace basic table with DataTableWrapper component using DataTableProvider
- Sort organizations by creation date (newest first) with default DESC order
- Maintain all existing functionality (badges, actions, delete confirmation)
- Add proper column definitions for DataTable with organization details
- Support configurable page sizes with default of 25 organizations per page
- Implement proper TypeScript types for pagination parameters
